updated error

+++ 14-binary_tree_balance.c
@@ -38,7 +38,7 @@ size_t binary_tree_height(const binary_tree_t *tree)
  * @tree: The binary tree
  * Return: The bf
  */
-int binary_tree_balance(const binary_tree_t *tree)
+size_t binary_tree_balance(const binary_tree_t *tree)
 {
 	if (tree == NULL)
 		return (0);
